# swagger.json
{
	"basePath": "/api",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"IBrand": {
			"description": "",
			"properties": {
				"id": {
					"type": "string",
					"description": ""
				},
				"name": {
					"type": "string",
					"description": ""
				},
				"devices": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"url": {
					"type": "string",
					"description": ""
				}
			},
			"required": [
				"id",
				"name",
				"devices",
				"url"
			],
			"type": "object"
		},
		"ICreateBrand": {
			"description": "",
			"properties": {
				"id": {
					"type": "string",
					"description": ""
				},
				"name": {
					"type": "string",
					"description": ""
				},
				"devices": {
					"type": "number",
					"format": "double",
					"description": ""
				},
				"url": {
					"type": "string",
					"description": ""
				}
			},
			"required": [
				"name",
				"url"
			],
			"type": "object"
		},
		"IUpdateBrand": {
			"description": "",
			"properties": {
				"id": {
					"type": "string",
					"description": ""
				},
				"name": {
					"type": "string",
					"description": ""
				},
				"devices": {
					"type": "string",
					"description": ""
				},
				"url": {
					"type": "string",
					"description": ""
				}
			},
			"required": [
				"id",
				"url"
			],
			"type": "object"
		},
		"IDevice": {
			"description": "",
			"properties": {
				"id": {
					"type": "string",
					"description": ""
				},
				"name": {
					"type": "string",
					"description": ""
				},
				"model": {
					"type": "string",
					"description": ""
				},
				"brand": {
					"type": "string",
					"description": ""
				}
			},
			"required": [
				"id",
				"name",
				"model",
				"brand"
			],
			"type": "object"
		},
		"ICreateDevice": {
			"description": "",
			"properties": {
				"id": {
					"type": "string",
					"description": ""
				},
				"name": {
					"type": "string",
					"description": ""
				},
				"model": {
					"type": "string",
					"description": ""
				},
				"brand": {
					"type": "string",
					"description": ""
				}
			},
			"required": [
				"name",
				"model",
				"brand"
			],
			"type": "object"
		},
		"IUpdateDevice": {
			"description": "",
			"properties": {
				"id": {
					"type": "string",
					"description": ""
				},
				"name": {
					"type": "string",
					"description": ""
				},
				"model": {
					"type": "string",
					"description": ""
				},
				"brand": {
					"type": "string",
					"description": ""
				}
			},
			"required": [
				"id"
			],
			"type": "object"
		}
	},
	"info": {
		"description": "blog-engine backend api endpoints documentations",
		"license": {
			"name": "MIT"
		},
		"title": "blog-engine",
		"version": "0.1.14"
	},
	"paths": {
		"/brands/": {
			"post": {
				"operationId": "BrandCreate",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IBrand"
						}
					}
				},
				"description": "",
				"summary": "",
				"tags": [
					"BrandController"
				],
				"parameters": [
					{
						"description": "",
						"in": "body",
						"name": "brand",
						"required": true,
						"schema": {
							"$ref": "#/definitions/ICreateBrand"
						}
					}
				]
			},
			"get": {
				"operationId": "BrandFindAll",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/IBrand"
							}
						}
					}
				},
				"description": "",
				"summary": "",
				"tags": [
					"BrandController"
				],
				"parameters": []
			}
		},
		"/brands/{id}": {
			"patch": {
				"operationId": "BrandUpdate",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IBrand"
						}
					}
				},
				"description": "",
				"summary": "",
				"tags": [
					"BrandController"
				],
				"parameters": [
					{
						"description": "",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					},
					{
						"description": "",
						"in": "body",
						"name": "brand",
						"required": true,
						"schema": {
							"$ref": "#/definitions/IUpdateBrand"
						}
					}
				]
			},
			"delete": {
				"operationId": "BrandDelete",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IBrand"
						}
					}
				},
				"description": "",
				"summary": "",
				"tags": [
					"BrandController"
				],
				"parameters": []
			},
			"get": {
				"operationId": "BrandFindOneById",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IBrand"
						}
					}
				},
				"description": "",
				"summary": "",
				"tags": [
					"BrandController"
				],
				"parameters": [
					{
						"description": "",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/devices/": {
			"post": {
				"operationId": "DeviceCreate",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IDevice"
						}
					}
				},
				"description": "",
				"summary": "",
				"tags": [
					"DeviceController"
				],
				"parameters": [
					{
						"description": "",
						"in": "body",
						"name": "device",
						"required": true,
						"schema": {
							"$ref": "#/definitions/ICreateDevice"
						}
					}
				]
			},
			"get": {
				"operationId": "DeviceFindAll",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/IDevice"
							}
						}
					}
				},
				"description": "",
				"summary": "",
				"tags": [
					"DeviceController"
				],
				"parameters": []
			}
		},
		"/devices/{id}": {
			"patch": {
				"operationId": "DeviceUpdate",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IDevice"
						}
					}
				},
				"description": "",
				"summary": "",
				"tags": [
					"DeviceController"
				],
				"parameters": [
					{
						"description": "",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					},
					{
						"description": "",
						"in": "body",
						"name": "device",
						"required": true,
						"schema": {
							"$ref": "#/definitions/IUpdateDevice"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeviceDelete",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IDevice"
						}
					}
				},
				"description": "",
				"summary": "",
				"tags": [
					"DeviceController"
				],
				"parameters": []
			},
			"get": {
				"operationId": "DeviceFindOneById",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/IDevice"
						}
					}
				},
				"description": "",
				"summary": "",
				"tags": [
					"DeviceController"
				],
				"parameters": [
					{
						"description": "",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {
		"api_key": {
			"type": "apiKey",
			"name": "access_token",
			"in": "query"
		},
		"tsoa_auth": {
			"type": "oauth2",
			"authorizationUrl": "http://swagger.io/api/oauth/dialog",
			"flow": "implicit",
			"scopes": {
				"write:pets": "modify things",
				"read:pets": "read things"
			}
		}
	},
	"host": "dev.localhost",
	"api_key": {
		"type": "apiKey",
		"name": "api_key",
		"in": "header"
	},
	"tsoa_auth": {
		"type": "oauth2",
		"authorizationUrl": "http://swagger.io/api/oauth/dialog",
		"flow": "implicit",
		"scopes": {
			"write:pets": "modify things",
			"read:pets": "read things"
		}
	}
}